refactor: data structure (#223)
### Summary of Changes

In this fix, the data structure is reworked to fit the model discussed
in the requirements.

Also, the analysis is changed regarding the resolving of references. 
`FunctionScope` class now holds target, value and call nodes for the
function, so we can iterate over these for each function.
It should now be more efficient and easier to understand since the
changes to the data structures below were used.

Added the class `Reference` to represent a node that references a
`Symbol`. The `ReferenceNode` class was extended with two subclasses
`TargetReference` - representing a Symbol referencing another Symbol,
and `ValueReference` - representing a Reference referencing a Symbol.
Both classes store the referenced symbols in a list.
Reworked the `Reasons` class so it now holds the Symbols for the
variables written to/ read from.
This blocks the merge of #211!


<!-- Please provide a summary of changes in this pull request, ensuring
all changes are explained. -->

---------

Co-authored-by: megalinter-bot <129584137+megalinter-bot@users.noreply.github.com>